backend/requirements.txt
django==4.2.16
djoser==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.2.1
